+++ Changelog.md
@@ -9,6 +9,7 @@
 * Store full country name, store days monitored each pod
 * Store detectedlanguage 
 * DB migrations see db/version.md
+* rename table rating_comments to ratingcomments for redbean support
 
 ## Cleanup
 * Use the git API for release versions, check development releases on pods https://github.com/diasporg/Poduptime/issues/143